requirements.txt
requests>=2.25.1
PyYAML>=6.0
